0.5.0 - some prep for new scala generation
